requirements.txt
# ADFLOWAI - Requirements
# Cleaned and tested - all packages install successfully on Python 3.11

# ── Core Framework ──────────────────────────────────────
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
flask-jwt-extended==4.6.0
gunicorn==21.2.0

# ── Database ─────────────────────────────────────────────
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# ── Cache / Queue ────────────────────────────────────────
redis==5.0.1
celery==5.3.4
flower==2.0.1

# ── Machine Learning (lightweight) ──────────────────────
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
joblib==1.3.2

# ── HTTP / Integrations ──────────────────────────────────
requests==2.31.0
requests-oauthlib==1.3.1

# ── Security ─────────────────────────────────────────────
bcrypt==4.1.2
pyjwt==2.8.0
cryptography==41.0.7

# ── Utilities ────────────────────────────────────────────
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
click==8.1.7

# ── Logging / Monitoring ─────────────────────────────────
prometheus-client==0.19.0
python-json-logger==2.0.7
